Refactor : 고객센터 초기 번들 크기 최적화 작업
